+++ src/OpenSearch.Client/Indices/IndexSettings/ComposableIndexTemplates/GetComposableIndexTemplate/GetComposableIndexTemplateRequest.cs
@@ -1,0 +1,15 @@
+/* SPDX-License-Identifier: Apache-2.0
+*
+* The OpenSearch Contributors require contributions made to
+* this file be licensed under the Apache-2.0 license or a
+* compatible open source license.
+*/
+
+namespace OpenSearch.Client;
+
+[MapsApi("indices.get_index_template")]
+public partial interface IGetComposableIndexTemplateRequest { }
+
+public partial class GetComposableIndexTemplateRequest { }
+
+public partial class GetComposableIndexTemplateDescriptor { }