+++ src/OpenSearch.Net/_Generated/IOpenSearchLowLevelClient.cs
@@ -50,6 +50,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using OpenSearch.Net;
+using OpenSearch.Net.Specification.ClusterApi;
+using OpenSearch.Net.Specification.IndicesApi;
 
 namespace OpenSearch.Net
 {
@@ -58,6 +60,12 @@ namespace OpenSearch.Net
     /// </summary>
     public partial interface IOpenSearchLowLevelClient
     {
+        /// <summary>Cluster APIs</summary>
+        LowLevelClusterNamespace Cluster { get; }
+
+        /// <summary>Indices APIs</summary>
+        LowLevelIndicesNamespace Indices { get; }
+
         /// <summary>POST on /{index}/_search/point_in_time <para>https://opensearch.org/docs/latest/search-plugins/point-in-time-api/#create-a-pit</para></summary>
         /// <param name="index">Comma-separated list of indices; use the special string `_all` or Indices.All to perform the operation on all indices.</param>
         /// <param name="requestParameters">Request specific configuration such as querystring parameters &amp; request specific connection settings.</param>